move decorators in the correct files
